requirements.txt
streamlit>=1.28.0
scikit-learn>=1.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pymupdf>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-pptx>=0.6.23
pypdf>=3.17.0